+++ HISTORY.rst
@@ -4,6 +4,8 @@ History
 
 * Add fixers to remove of various compatibility imports removed in Django 3.1.
 
+* Add fixer for Django 2.2 to rewrite ``request.META`` access of headers to ``HttpRequest.headers``.
+
 1.2.0 (2021-09-02)
 ------------------